refactor(store): move the runtime monitor and boot recovery out of store-health-check
store-health-check.ts had absorbed the 2026-07-18 survivability build's
operational machinery — probe timers, docker restart orchestration,
harden-lock suspension, boot-recovery cooldown persistence — on top of
its original job (pure reachability probes, namespace identity tagging,
operator-facing formatting). A module named for health CHECKS owning
docker restarts and lock files is exactly the boundary drift that makes
both halves harder to reason about.

New store-runtime-monitor.ts owns createStoreRuntimeMonitor,
attemptManagedStoreBootRecovery, and the harden-lock / boot-restart-ts
path helpers (the lock helpers move WITH the monitor because the
monitor is the lock's reader — keeping reader and path-definition
together is the original design intent, now with the right module
name). It depends on store-health-check's checkExternalStoreReachable
for probing; nothing flows the other way.

Pure move — no logic change. Importers updated (lifecycle, daemon
state, harden executor, tests); store-health-check.ts is back to
probe + identity + formatting only, and its fs/docker imports are gone.

Review thread: store-health-check.ts:565 (keep Docker recovery
orchestration out of the health-check module).
